scenarios/iot_device_provisioning/pyproject.toml
[tool.poetry]
name = "iot-device-provisioning"
version = "0.0.1"
description = "provisioning IoT devices"
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
python-dotenv = "^1.0.1"
azure-iot-device = "^2.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
